requirements.txt
# Build Tools (install first)
setuptools>=65.0.0
wheel

# Web Frameworks
fastapi>=0.115.2
uvicorn>=0.20.0
starlette>=0.40.0

# UI Dashboard
streamlit>=1.28.0

# Core Utilities
requests>=2.32.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.7.4
python-dotenv>=1.0.0

# Database
# Note: sqlite3 is included in Python standard library - no package needed

# AI / NLP / LLM
google-generativeai>=0.3.0
openai>=1.86.0

# Security APIs and Tools
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Data Processing
PyYAML>=6.0.0
markdown>=3.5.0

# Visualization and Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
reportlab>=4.0.0

# Date and Time
python-dateutil>=2.8.0

# Logging and Monitoring
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
